pyproject.toml
[project]
name = "lithium"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "django ~=5.1",
  "django-allauth[openid,socialaccount] ~=65.2",
  "django-debug-toolbar ~=4.4",
  "crispy-bootstrap5 ~=2024.10",
  "gunicorn ~=23.0",
  "psycopg[binary] ~=3.2",
  "whitenoise ~=6.7",
  "langchain-openai>=1.1.7",
  "langchain-core>=1.2.8",
  "langchain>=1.2.8",
  "python-dotenv>=1.2.1",
  "django-widget-tweaks>=1.5.1",
  "django-markdownify>=0.9.6",
  "environs",
  "django-cache-memoize>=0.2.1",
  "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
  "typing_extensions ~=4.12",
]
